Adjust SP-based offsets passing as a function argument
All stack addresses in SyncVM are SP based. When a function takes a
stack offset as a parameter, it has to add the size of its frame to it.
The patch introduces a draft mechanism implementing the aforementioned
logic. When lowering formal parameters it instruments all stack (address
space 0) pointers with AdjSP pseudo instruction. Later
SyncVMAdjustSPBasedOffsets pass adjust offsets of loads and stores using
AdjSP as an input.
The patch also fixes push and pop instructions when creating a function's
frame.