Support unaligned offsets when addressing memory
In case a displacement is not divisible by 32, add displacement % 32 to
the base register and use the rest as the displacement itself. Thus it
can be represented in cells.
In case there is no base register, use the closest divisible by 32
offset less then the displacement.
The patch also replaces i16 to i64 for displacement handling.