Fix BB placement issue
Move call spill result pass used to insert a BB so that the previous BB
fall through and control flow unexpectedly reach the inserted BB. The
patch places a newly inserted BB so that it follows the original block
that spills. Since it has to have EH logic, the jlt terminator is
guaranteed.