Adding SyncVMLowerInvokes pass to handle InvokeInst
This patch deals with exception handling on `InvokeInst`s.
Basically we will convert it into a regular call and some checking
branches after returning from the call. a global variable is created
to serve as the indicator of the sign of exception.

What has not been handled are the types of exception, we could extend
the exception to take signals in the future.