Fix returndata not saved in THROW issue
After wrapping the THROW into a wrapper call, we artifically adds
a new local frame with no unwind path, which will result in lost
of register values in case of exception.

The fix is to directly inline calls to the wrapper call. In this case
we just simply replace the call with THROW.