[EraVM] Propagate constant to a stack accessing instruction (#84)
Propagate constants to stack address

In case a fuction access to an external frame outside of the entry basic
block, the DAG is constructed so that the address is computed in entry
and copied to a register. SelectAddress therefore can't infer that a
constant could be involved in the computation and use Reg + Imm
addressing mode.
The patch introduces a pass that replicates the logic of propagation
constants to a memory accessing operation. If the inital address is
computed by expression 1, and none of it's patrital results is used more
than once, the pass attempts to find expression 2, so that expression 1
is equal to expression 2 + constant. Then it replaces a memory access by
expression 1 to a memory access by expression 2 + constant.