[EraVM] Update memcpy expansion
Expand memcpy using gep instead of pointer arithrmetic to access
elements, and move the logic to EraVM target instead of making local
changes in LowerMemIntrinsics.
The patch also corrects the data layout for EraVM.

Do not optimize memcpy if size is bigger than i64. On EraVM we support
i256 type sizes, hence much bigger possible value space.
So we have to check and make sure the modified LLVM pipeline is not going to
crash should a very large size comes in.

We simply bail out optimizations on cases that the datatype is too big to handle.