Handle spilling of fat pointer in R1
If to use Greedy register allocator, the only source of physical
registers by the time spills and reloads are generated is in calling
conventions lowering. The patch recognize result of a far call passed in
R1 as a fat pointer and uses ptr.add for spills and reloads of it.