[EraVM] Introduce 'MVT::fatptr' datatype for generic AS pointers (#127)
Fat pointer type is to used to distinguish with regular i256 type,
which does not require special instruction handling to maintain its
tagged status within VM.

Pointers to generic address space will be interpreted as fat pointers
in the backend. It will be assigned a physical register which could be
overlapped with i256 -- this is conformant to fat pointer semantics.