[ISEL] Redo stack address selection
Existing stack address selection mechanism is sometimes hard to follow.
This patch tries to re-model its behavior to make it easier to understand
and debug.

This patch still uses 3 operands to represent a stack operand:
* The first operand is Base, it must be a register.
  $r0 is allowed in relative addressing mode.
* The second operand is Disp, it must be immediate or expression.
* the third operand is isRelative, which is a boolean.

Note that we might later change the third operand which will allow
us to represent indexed stack operations such as `stack-=[1]`.