Improve runtime linkage
Split runtime in two parts: runtime and stdlib. Stdlib is the part to be
linked and internalized in the beginning of opt pipeline. Then global
DCE removes the unused parts. So small contracts doesn't blow up and
affect compilation time. Runtime is linked in the beginning of llc
pipeline.

This solution replaces previously used internalize-early,
internalize-late function attributes.