Do not optimize memcpy if size is bigger than i64
On SyncVM we support i256 type sizes, hence much bigger possible value space.
So we have to check and make sure the modified LLVM pipeline is not going to
crash should a very large size comes in.

We simply bail out optimizations on cases that the datatype is too big to handle.