Fix one bug in SyncVMCombineFlagSettings
When trying to combine flag setting instructions, we should also merge
implicit defs.  However, the `implicit-def $flag` was missing in the
combined instruction.

This will cause LICM or instruction sinking thinking that the combined
instructions are dep-free and hence can be moved freely. It should have shown
in tests previously but was not.

Also adds a few additional tests to explicitly check that the combined
instruction has an implicit def of $flags.