[EVM] Addign frameindex lowering.
Please, note this implementation assumes the stack is located in
AS0 memeory and is intended for getting MIR with virtual register.
We need to design how the frame slots to be represented in the stackified code.
We can either use EVM stack (though EVM doesn't seem to have instructions to
do this in a general case), or we can allocate a stack area in the AS0 memory.