[SyncVM] Moving FE runtime library functionality to the BE.
Added implementation of 'div', 'sdiv', 'mod', 'smod', 'shl', 'shr', 'sar',
'byte', 'mstore8' and sha3 operations to the syncvm-stdlib.ll along with
the support of constant folding.

Stdlib calls with at least one 'Undef'/'Poision' argument are folded to 'Poison'
value. 'Undef'/'Poison' values are not expected among input values in stdlib
calls. Presence of such values may denote an invalid LLVM IR. This should help
in revealing invalid IR, by inducing UB in the subsequent passes.

SyncVMLinkRuntime pass was extended to set 'noinline', 'minsize' and
'optforsize' attributes to functions of a module on 'Oz' mode.
The goal was to mimic FE original behavior. The FE is setting 'minsize' and
'optforsize' for literally all functions on 'Mz' mode. 'noinline' is only set
for some runtime/stdlib functions. According to the experiments, inlining them
leads to a size explosion thus mitigating the purpose of the Mz mode.
'noinline' attribute is set if the function definition has the 'noinline-oz'
target attribute that was especially introduced.

SyncVMLinkRuntime pass has been moved from the beginning of the pipeline
to the point after some basic optimizations. This enables more refinement
selection of the functions to be inlined and which to be not.