[RDA] Fix getGlobalUses() to account for redefinitions
Prior to the patch getGlobalUses() continued to look for uses in
a successor BB if the register of interest is alive in the successor,
but it didn't check whether the BB redefined the register.
The patch adds this check, so getGlobalUses() only proceed with successors
if the register is not redefined.

Upstream PR: https://github.com/llvm/llvm-project/pull/78178