[EraVM][AsmParser] First cut of instruction parsing
Remove outdated stuff, add support for simple cases of instruction parsing:
* predicates and arithmetic modifiers
* register and immediate operands
* code and absolute stack operands

At now, there are several special cases in how the backend generates
instruction operands to communicate them to asm printer - move the logic
for adding the expected operands to a separate appendMCOperands() function
which is called by code in the asm parser. For that purpose, I assume it
is safe to store pointer to MCContext inside an EraVMOperand field.