[EraVM] Don't traverse over subloop BBs in EraVMIndexedMemOpsPrepare
This fixes the crash where inner loop is not a simplified form and
we are traversing over the basic blocks of the outer loop and
attempt to rewrite the inner GEP instruction using the preheader
and latch basic blocks of the outer loop. Additionally, this code
adds a paranoia check to ensure that the PHI node of the BasePtr
has the preheader and latch basic blocks as incoming blocks.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>