[EraVM][AsmPrinter] Reuse EraVM::analyzeMCOperandsCode/Stack functions.
Reuse the same `parseMCOperandsCode` and `parseMCOperandsStack`
functions that are used by code emitter to analyze operands of MCInst.

While it makes easier to improve formatting of printed assembly code,
this commit strives to keep the existing formatting as much as possible,
to not combine refactoring with massive changes of tests.