[EraVM][MC] Refine the definitions of ret-like instructions
Change return and revert instructions to either accept jump target or
register operand, but not both at the same time.

Change panic instruction to either accept jump target operand
or none at all.

Hardcode the value of rs0 field used to encode the instruction variants
with jump target operand: 1 for return and revert and 0 for panic.