[EraVM] Optimize overflow intrinsic which has a branching use
For every branch that utilizes the overflow i1 output of uaddo, usubo,
umulo, the FBB is moved next to the branch in layout and the TBB is
moved out of the way to the cold section. Thus the code generator can
use `jump.lt` to jump to TBB and reach FBB without a jump. We expect
that overflow is a less probable scenario for a typical usage.
To implement the described logic, we lower `BRCOND` node to
a conditional jump and glue it to the flag setting add, sub or mul.

Co-authored-by: Alan Li <alan.li@me.com>
PR: #581, Issue: #491.