[EraVM] Enable main induction variable removal in LSR
This patch makes `shouldFoldTerminatingConditionAfterLSR` return `true`
for EraVM target. Thus LSR will try to replace main induction variable
of a loop with one of secondary IVs. Thus it reduces the number of
instructions in a loop by eliminating main IV increment.
This transformation increases register pressure, but it's rarely the
problem for EraVM.

PR: #599, Issue: #580.