[EraVM] Change lowering of memmove
This patch improves lowering of memmove by reusing a gep
instruction to calculate the end dst address of copy forward
part in the loop and the residual BBs. To do so, it removes
the main IV in the expanded loop and replaces its
condition with a comparison of dst pointer with end dst
pointer. The same end dst pointer is reused in the
residual BB. Normally the loop transformation is to be
done by LSR pass but as we want to use dst pointer in
the residual BB, we rewrite the IR for lowering memory
intrinsics beforehand.

PR: #633.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>