[EraVM] Adjust CMP to expose CSE opportunities with SUB
This patch adds a pass that adjusts CMP with immediate
instructions to expose CSE opportunities with SUB. To do
so, it needs to adjust immediate in CMP instruction and
to adjust conditional code in the next instruction that
should be a select or a unconditional branch.
As a result it changes following assembly:
```
  sub.s   10, r1, r2
  sub.s!   9, r1, r0
  jump.gt @.BB0_2
```
into
```
  sub.s   10, r1, r2
  sub.s!  10, r1, r0
  jump.ge @.BB0_2
```
So MachineCSE can eliminate redundant sub.

PR: #651.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>