[MC][AsmLexer] 'LexToken()': fix potential buffer overflow.
When the 'CurPtr' points to the 'EOF'
(which corresponds to 'CurBuf.end()'), calling the
'isAtStartOfComment'/'isAtStartOfComment' function leads to
dereferencing of 'CurBuf.end()'.
Usually this issue is hidden, because the AsmParser receives a source
code via MemoryBuffer object with the null-terminating symbol, but the
null-terminator is not required for AsmParser logic.