[EraVM] Add the disassembler C-API.
This API accepts just a bare EraVM byte code (not in the ELF format).
As a result, we employ heuristic-base approach to understand the byte code
structure. In particular, to understand where a section with constants begins,
is the given byte octet an instruction, padding, or metadata.
This does not work in general case, so a user may need to check hex encoding
and consult with the EraVM binary layout, to fully understand the byte code
structure.