[APInt] Replace an assertion in constructor with fatal error
The patch replaces assertions in APInt constructor with
`report_fatal_error` calls.
EraVM compiler is shipped to users w.o. assertions enabled, but zero
extensions of negative numbers is a strong signal that a transformation
which called the constructor has a bug. In that case we prefer to fail
with diagnostic instead of miscompiling a module even if false
positives could happen.