[TailDuplication] Add option to enable tail duplication of fallthrough BBs
During the TailDuplication pass, fallthrough BBs are
not duplicated. This is generally acceptable, but in
scenarios where a loop contains a large switch case
and the latch BB is a fallthrough, duplicating the
latch BB can eliminate jump instructions in its
predecessors. This patch introduces the
`tail-dup-fallthrough-bbs` option to enable the
duplication of fallthrough BBs.

PR: #727.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>