[EVM] Remove Exits from FunctionInfo
This commit removes Exits data member from
FunctionInfo struct, but it adds Exits to
ControlFlowGraphBuilder just to reduce number
of changes needed to mark blocks that needs
to maintain a clean stack. It also fixes comment
in markNeedsCleanStack as although we are calling
AddChild, we are adding predecessors (parents) and
not successors (childs).

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>