[EraVM] Add support for sinking loads to their users
This optimization sinks load instructions to their users,
if there are no instructions that may clobber the memory
between them. This is useful, since loads in some cases
can be combined with their users to reduce the number of
instructions. Currently, this optimization is not enabled
by default and it is limited to loads from loop header that
load from a global value that holds a pointer to the generic
address space, and if all users are addp instructions.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>