[EVM] Support evmla pushdeployaddress instruction
This implements 'Call Protection For Libraries' mechanism in EVMLA mode.

The runtime code of a library always starts with a push instruction,
which is a zero of 20 bytes at compilation time. When the deploy code runs,
this constant is replaced in memory by the current address and this modified
code is stored in the contract. The front-end handles the deploy-related part.
At runtime, this causes the deploy time address to be the first constant to be
pushed onto the stack and the dispatcher code compares the current address
against this constant for any non-view and non-pure function.

TODO: #778. The correctness of this patch is based on the empirical observation
that the function containing PUSHDEPLOYADDRESS always appears first in the
module layout. Theoretically, this may not hold, so we need to position this
function at the start.