[EVM] Disable tail duplication optimization
In some cases, tail duplication can combine a loop’s
header with its latch. When this happens,
MachineLoopInfo may fail to recognize the loop during
stackification. This is a problem because, in
EVMStackSolver::runPropagation, we add an empty exit
stack for loop latches during the first run (if the
header hasn’t been visited yet) to prevent infinite
loops. If the loop isn’t detected, we will run into
an infinite loop situation.

This is a workaround and #822 should fix it in a proper
way.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>