[EVM] Introduce APIs for retrieving spill area size
This patch introduces functionality to inform the front-end (FE)
about the spill area size determined by the stackifier. It includes:

A new C-API for the FE to query the spill area size:

  uint64_t LLVMGetSpillAreaSizeEVM(LLVMContextRef C);

New methods in LLVMContext for setting and retrieving the spill area
size calculated by the stackifier:

  void setSpillAreaSize(uint64_t Size);

  // Internal use only, utilized by LLVMGetSpillAreaSizeEVM
  uint64_t getSpillAreaSize() const;

Spill/Reload workflow:

 - The FE compiles LLVM IR normally without specifying any spill-related
   options

 - If no stackification (STD) errors occur, compilation completes successfully
   and emits object code

 - If a STD error occurs:
    - The EVMFinalizeStackFrames pass sets the required spill area size via
      LLVMContext::setSpillAreaSize

    - Compilation fails gracefully (not via assertion);
      LLVMTargetMachineEmitToMemoryBuffer returns true

    - The FE then invokes LLVMGetSpillAreaSizeEVM to retrieve the required
      spill area size

    - The FE allocates the appropriate spill area and recompiles the LLVM IR,
      passing the allocated size via the StackRegionSize option and the memory
      start address via StackRegionOffset.

Note:
The StackRegionSize and StackRegionOffset options are used to pass spill area
information from the FE (or llc) to the EVM back-end. In contrast, this patch
provides an API for passing the required spill area size from the EVM back-end
to the FE.