[EVM] Introduce APIs for retrieving spill area size
This patch introduces functionality to inform the front-end (FE)
about the spill area size determined by the stackifier. It includes:

A new C-API for the FE to query the spill area size:

  uint64_t LLVMGetSpillAreaSizeEVM(LLVMContextRef C);

New methods in LLVMContext for setting and retrieving the spill area
size calculated by the stackifier:

  void setSpillAreaSize(uint64_t Size);

  // Internal use only, utilized by LLVMGetSpillAreaSizeEVM
  uint64_t getSpillAreaSize() const;

Spill/Reload workflow:

 - The FE compiles LLVM IR normally without specifying any spill-related
   options

 - If no stackification (STD) errors occur, compilation completes successfully
   and emits object code

 - If a STD error occurs:
    - The EVMFinalizeStackFrames pass sets the required spill area size via
      LLVMContext::setSpillAreaSize

    - Compilation fails gracefully (not via assertion);
      LLVMTargetMachineEmitToMemoryBuffer returns true

    - The FE then invokes LLVMGetSpillAreaSizeEVM to retrieve the required
      spill area size

    - The FE allocates the appropriate spill area and recompiles the LLVM IR,
      passing the allocated size via the StackRegionSize option and the memory
      start address via StackRegionOffset.

Note:
The StackRegionSize and StackRegionOffset options are used to pass spill area
information from the FE (or llc) to the EVM back-end. In contrast, this patch
provides an API for passing the required spill area size from the EVM back-end
to the FE.

This patch could be simplified—or potentially avoided altogether if a
callback-based mechanism were available for communicating with the front-end.
The idea would look like this:

1. Defining a stack-too-deep handler in Rust:

  void on_std(uint64_t SpillAreaSize) {
    ...
    exit(0);
  }

2. Registering the callback in LLVM:

  LLVMInstallEVMStdHandler(on_std);

3. Invoking the handler from the EVM backend:

  report_evm_std(32);

With this approach, there would be no need to modify LLVMContext or implement
custom logic to interrupt the compilation pipeline. However, it introduces
added complexity on the front-end side specifically, implementing the callback
in Rust and ensuring clean termination of the current process or thread when a
stack-too-deep (STD) error occurs. Given these trade-offs, this approach was
ultimately not pursued.