[EVM] Mark recursive functions with attribute
Find all recursive functions in a module and mark
them with 'evm-recursive' attribute. This attribute
is used later during stackification to issue an error
if we find stack too deep errors, since we can't use
spills in that case. Reason why we can't is that for
spills we are using memory and not the real stack.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>