[CGP] Disable optimizeMemoryInst only for EraVM
This enables optimization for EVM and implements
isLegalAddressingMode, allowing r + imm addressing
mode for all address spaces except CALLDATALOAD.
Benchmarks indicate that enabling it for CALLDATALOAD
is not beneficial. This change primarily allows
address mode calculations (r + imm) to be sunk
to their uses, which can reduce spills and
reloads in some cases.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>