[EVM] Clear the branch condition in TII::removeBranch for stackified code
The issue is that TII::removeBranch removes only the branch instruction
itself, leaving behind the code that computes the branch condition which
breaks the stack layout. To address this, we add a POP instruction after
removing the branch in TII::removeBranch to clear the top of the stack.
In TII::insertBranch, we remove the POP instruction since the condition
will be consumed by the newly inserted conditional branch.