[EVM] Add support for CLZ instruction
CLZ will be introduced in the fusaka update:
https://eips.ethereum.org/EIPS/eip-7939

Beside adding support for CLZ instruction in this
patch, support for EVM versioning is added, so CLZ
instruction is generated for fusaka and later versions.
isCheapToSpeculateCttz and isCheapToSpeculateCtlz
are implemented so CGP doesn't generate check for zero
to avoid calling the intrinsic in despeculateCountZeros.
CTZ is generated via CLZ, and currently LLVM is
generating:
`256 - clz(~x & (x-1))`
If x != 0 is proven, this can be optimized to:
`255 - clz(x & -x)` (#918)

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>