[EVM] Fix nondeterminism in constant unfolding
Previously, the code that collected PUSH instructions
to process used a DenseMap with MachineFunction* as
the key. This could lead to nondeterministic iteration
order, causing inconsistent behavior across different
runs. Fix this by changing DenseMap to SmallVector,
since we are using lookups for asserts that can't fail,
because we are iterating over all MachineFunctions
only once.

Signed-off-by: Vladimir Radosavljevic <vr@matterlabs.dev>